// src/pages/storefront-designer/types/designerTypes.ts
export interface SocialPreview {
  key: string;
  value: string;
  showPreview: boolean;
}

/**
 * Additional links configuration
 */
export interface AdditionalLinks {
  caption: string;
  link: string;
  save: boolean;
}

/**
 * Shop design configuration
 */
export interface ShopDesignPage {
  templateID?: string;
  backgroundImage?: string;
  backgroundImageSecondary?: string;
  fullWidthHero?: boolean;
  productSectionText?: string;
  backgroundColor?: string;
  backgroundText?: string;
  logo?: string;
  discordURL?: string;
  instagramURL?: string;
  twitterURL?: string;
  facebookURL?: string;
  linkedinURL?: string;
  tiktokURL?: string;
  webURL?: string;
  telegramURL?: string;
  youtubeURL?: string;
  messengerURL?: string;
  headerIcon?: string;
  shopDesign?: {
    fontfamily?: string;
    headerBackground?: string;
    isHeaderFixed?: boolean;
    hiroLayout?: string;
    hiroTextColor?: string;
    bannerLinks?: Array<AdditionalLinks>;
    footerLinks?: Array<AdditionalLinks>;
    productListTitle?: string;
    iconHeaderColor?: string;
    foreground?: string;
    backgroundBody?: string;
    textColorParagraphs?: string;
    isCollectionShown?: boolean;
    isLogoAsFavicon?: boolean;
    faviconURL?: string;
  };
  template_options?: ITemplateOptions;
  launchDate?: string | null;
}

/**
 * Designer page state
 */
export interface StateDesignPage {
  device?: 'desktop' | 'tablet' | 'mobile';
  shop?: ShopDesignPage;
  activeSection?: string;
  socialPreview?: SocialPreview;
}